# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/cms/schedule/{id}": {
      "get": {
        "operationId": "ScheduleCmsController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "cms/schedule"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "put": {
        "operationId": "ScheduleCmsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateScheduleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "cms/schedule"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "delete": {
        "operationId": "ScheduleCmsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "cms/schedule"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
...
    "/v1/public/planting-schedule": {
      "get": {
        "operationId": "PlantingScheduleController_list",
        "summary": "Daftar musim tanam publik",
        "description": "Mengembalikan musim tanam beserta tahapan-tahapannya. Default hanya status AKTIF. Filter opsional: ?komoditasId=&status=",
        "parameters": [
          {
            "name": "komoditasId",
            "required": false,
            "in": "query",
            "description": "Filter berdasarkan ID komoditas",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter status (default: AKTIF)",
            "schema": {
              "enum": [
                "DRAFT",
                "AKTIF",
                "SELESAI"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "public/planting-schedule"
        ]
      }
    },
...
    "/v1/public/planting-schedule/{id}": {
      "get": {
        "operationId": "PlantingScheduleController_findOne",
        "summary": "Detail musim tanam",
        "description": "Detail satu musim tanam beserta seluruh tahapan-tahapannya, diurutkan berdasarkan field urutan.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "public/planting-schedule"
        ]
      }
    },
    "/v1/cms/planting-schedule": {
      "get": {
        "operationId": "PlantingScheduleCmsController_listAll",
        "summary": "Semua musim tanam (admin)",
        "description": "Mengembalikan semua musim tanam tanpa filter status. Filter opsional: ?komoditasId=&status=",
        "parameters": [
          {
            "name": "komoditasId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "DRAFT",
                "AKTIF",
                "SELESAI"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "cms/planting-schedule"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "post": {
        "operationId": "PlantingScheduleCmsController_create",
        "summary": "Buat musim tanam baru",
        "description": "Membuat musim tanam baru. Opsional: sertakan array `tahapan` untuk membuat tahapan awal sekaligus. createdById diambil otomatis dari token JWT (tidak perlu di body).",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMusimTanamDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "cms/planting-schedule"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/v1/cms/planting-schedule/{id}": {
      "get": {
        "operationId": "PlantingScheduleCmsController_findOne",
        "summary": "Detail musim tanam (admin)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "cms/planting-schedule"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "put": {
        "operationId": "PlantingScheduleCmsController_update",
        "summary": "Update musim tanam",
        "description": "Update data header musim tanam (judul, lokasi, status, dll). Untuk update tahapan gunakan endpoint /tahapan.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMusimTanamDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "cms/planting-schedule"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "delete": {
        "operationId": "PlantingScheduleCmsController_remove",
        "summary": "Hapus musim tanam",
        "description": "Hard delete musim tanam beserta seluruh tahapan-tahapannya (cascade).",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "cms/planting-schedule"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/v1/cms/planting-schedule/{musimTanamId}/tahapan": {
      "post": {
        "operationId": "PlantingScheduleCmsController_createTahapan",
        "summary": "Tambah tahapan ke musim tanam",
        "description": "Menambahkan satu tahapan baru ke musim tanam yang sudah ada.",
        "parameters": [
          {
            "name": "musimTanamId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTahapanDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "cms/planting-schedule"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/v1/cms/planting-schedule/tahapan/{id}": {
      "put": {
        "operationId": "PlantingScheduleCmsController_updateTahapan",
        "summary": "Update tahapan jadwal",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTahapanDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "cms/planting-schedule"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "delete": {
        "operationId": "PlantingScheduleCmsController_removeTahapan",
        "summary": "Hapus tahapan jadwal",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "cms/planting-schedule"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/v1/cms/planting-schedule/komoditas": {
      "post": {
        "operationId": "PlantingScheduleCmsController_createKomoditas",
        "summary": "Buat komoditas baru",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateKomoditasDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "cms/planting-schedule"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/v1/cms/planting-schedule/komoditas/{id}": {
      "put": {
        "operationId": "PlantingScheduleCmsController_updateKomoditas",
        "summary": "Update komoditas",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateKomoditasDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "cms/planting-schedule"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "delete": {
        "operationId": "PlantingScheduleCmsController_removeKomoditas",
        "summary": "Hapus komoditas",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "cms/planting-schedule"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "NAWASENA Center API",
    "description": "Backend NAWASENA Center untuk PPK Ormawa RDC UAD.\n\n## Cara pakai\n- Endpoint **public/** bisa diakses tanpa login (untuk website publik).\n- Endpoint **cms/** butuh login admin. Klik tombol **Authorize** di kanan atas, masukkan token (format: `Bearer <token>` tidak perlu, cukup token-nya).\n\n## Format response\nSemua response: `{ success: true, data: {...} }` atau `{ success: false, error: {...} }`.\n\n## Autentikasi\nLogin via Supabase Auth untuk dapat access_token, lalu Authorize di sini.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Masukkan access_token dari Supabase Auth (tanpa kata Bearer)"
      }
    },
    "schemas": {
...
      "UpdateScheduleDto": {
        "type": "object",
        "properties": {
          "farmerName": {
            "type": "string",
            "example": "Pak Sutrisno",
            "description": "Nama petani"
          },
          "fieldName": {
            "type": "string",
            "example": "Sawah Blok A",
            "description": "Nama/kode lahan"
          },
          "padukuhan": {
            "type": "string",
            "example": "Karangtengah",
            "description": "Padukuhan (untuk tanam serentak per wilayah)"
          },
          "plantDate": {
            "type": "string",
            "example": "2026-06-20",
            "description": "Tanggal tanam (YYYY-MM-DD)"
          },
          "estimatedHarvest": {
            "type": "string",
            "example": "2026-10-13",
            "description": "Perkiraan panen. Kosongkan untuk hitung otomatis dari umur padi."
          },
          "cropDays": {
            "type": "number",
            "example": 115,
            "description": "Umur padi (hari) untuk hitung estimasi panen otomatis. Default 115."
          },
          "notes": {
            "type": "string",
            "example": "Varietas IR64"
          },
          "status": {
            "type": "string",
            "example": "planted",
            "description": "planned | planted | harvested"
          }
        }
      },
...
      "CreateTahapanDto": {
        "type": "object",
        "properties": {
          "namaTahapan": {
            "type": "string",
            "example": "Penyemaian Benih",
            "description": "Nama tahapan kegiatan"
          },
          "tanggalMulai": {
            "type": "string",
            "example": "2026-09-02T00:00:00.000Z",
            "description": "Tanggal mulai tahapan (ISO 8601)"
          },
          "tanggalSelesai": {
            "type": "string",
            "example": "2026-09-10T00:00:00.000Z",
            "description": "Tanggal selesai tahapan (opsional)"
          },
          "urutan": {
            "type": "number",
            "example": 1,
            "description": "Urutan tampil di timeline (mulai dari 1)"
          },
          "deskripsi": {
            "type": "string",
            "example": "Benih disemai di persemaian selama 21 hari"
          }
        },
        "required": [
          "namaTahapan",
          "tanggalMulai",
          "urutan"
        ]
      },
      "CreateMusimTanamDto": {
        "type": "object",
        "properties": {
          "judul": {
            "type": "string",
            "example": "Musim Tanam Gadu 2026",
            "description": "Judul musim tanam (wajib)"
          },
          "komoditasId": {
            "type": "string",
            "example": "uuid-komoditas",
            "description": "ID Komoditas (UUID)"
          },
          "lokasi": {
            "type": "string",
            "example": "Blok A - Karangtengah",
            "description": "Blok/wilayah sawah"
          },
          "tanggalMulai": {
            "type": "string",
            "example": "2026-09-01T00:00:00.000Z",
            "description": "Tanggal mulai musim tanam (ISO 8601)"
          },
          "tanggalSelesai": {
            "type": "string",
            "example": "2026-12-31T00:00:00.000Z",
            "description": "Tanggal selesai musim tanam (ISO 8601)"
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "AKTIF",
              "SELESAI"
            ],
            "default": "DRAFT",
            "description": "Status musim tanam: DRAFT | AKTIF | SELESAI"
          },
          "tahapan": {
            "description": "Daftar tahapan awal (opsional, bisa ditambah belakangan)",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateTahapanDto"
            }
          }
        },
        "required": [
          "judul",
          "komoditasId"
        ]
      },
      "UpdateMusimTanamDto": {
        "type": "object",
        "properties": {
          "judul": {
            "type": "string",
            "example": "Musim Tanam Gadu 2026",
            "description": "Judul musim tanam (wajib)"
          },
          "komoditasId": {
            "type": "string",
            "example": "uuid-komoditas",
            "description": "ID Komoditas (UUID)"
          },
          "lokasi": {
            "type": "string",
            "example": "Blok A - Karangtengah",
            "description": "Blok/wilayah sawah"
          },
          "tanggalMulai": {
            "type": "string",
            "example": "2026-09-01T00:00:00.000Z",
            "description": "Tanggal mulai musim tanam (ISO 8601)"
          },
          "tanggalSelesai": {
            "type": "string",
            "example": "2026-12-31T00:00:00.000Z",
            "description": "Tanggal selesai musim tanam (ISO 8601)"
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "AKTIF",
              "SELESAI"
            ],
            "default": "DRAFT",
            "description": "Status musim tanam: DRAFT | AKTIF | SELESAI"
          },
          "tahapan": {
            "description": "Daftar tahapan awal (opsional, bisa ditambah belakangan)",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateTahapanDto"
            }
          }
        }
      },
      "UpdateTahapanDto": {
        "type": "object",
        "properties": {
          "namaTahapan": {
            "type": "string",
            "example": "Penyemaian Benih",
            "description": "Nama tahapan kegiatan"
          },
          "tanggalMulai": {
            "type": "string",
            "example": "2026-09-02T00:00:00.000Z",
            "description": "Tanggal mulai tahapan (ISO 8601)"
          },
          "tanggalSelesai": {
            "type": "string",
            "example": "2026-09-10T00:00:00.000Z",
            "description": "Tanggal selesai tahapan (opsional)"
          },
          "urutan": {
            "type": "number",
            "example": 1,
            "description": "Urutan tampil di timeline (mulai dari 1)"
          },
          "deskripsi": {
            "type": "string",
            "example": "Benih disemai di persemaian selama 21 hari"
          }
        }
      },
      "CreateKomoditasDto": {
        "type": "object",
        "properties": {
          "nama": {
            "type": "string",
            "example": "Padi",
            "description": "Nama komoditas"
          },
          "deskripsi": {
            "type": "string",
            "example": "Komoditas utama padi sawah"
          }
        },
        "required": [
          "nama"
        ]
      }
    }
  }
}